ent-est-v3/services/users/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
cassandra-driver==3.29.0
pyjwt==2.8.0
bcrypt==4.1.2
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.6.4
requests==2.31.0
cryptography==42.0.5
